✨ Feat :  commentWrapper 컴포넌트로 분리
댓글 API 연동하기 위해 컴포넌트 분리하였습니다.

#63
